finish route migration
